+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- version information updated [#38](https://github.com/asphaltbuffet/ogma/issues/38)
+
+### Removed
+
+- completion command disabled [#37](https://github.com/asphaltbuffet/ogma/issues/37)
+
+### Fixes
+
+- base root command no longer returns an error [#35](https://github.com/asphaltbuffet/ogma/issues/35)
+
 ## [1.0.0] - 2021-11-15
 
 ### Added